Fix delegateAddress type
The Delegate EIP712 payload was using bytes32 as the type for the delegate address (which is an Ethereum Address).

While this works, it represents the wrong input type and size constraints that we should accept for that field (an address is 20 bytes so any input bigger than that should not be valid).